src/requirements.txt
# Multi-AI Agent Systems with CrewAI - Dependencies
# This file contains all the necessary packages to run the Python implementations

# Core CrewAI framework
crewai==0.28.8

# CrewAI tools for enhanced functionality
crewai-tools==0.1.6

# LangChain community tools and integrations
langchain-community==0.0.29

# Pydantic for data validation and settings
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Data handling and manipulation
pandas>=1.5.0
numpy>=1.21.0

# Web scraping and HTTP requests
requests>=2.28.0
beautifulsoup4>=4.11.0

# JSON and data processing
jsonschema>=4.0.0

# Optional: Jupyter support (if you want to run the original notebooks)
# jupyter>=1.0.0
# ipykernel>=6.0.0

# Optional: Additional LLM providers
# openai>=1.0.0
# anthropic>=0.7.0
# cohere>=4.0.0

# Optional: Vector databases and embeddings
# chromadb>=0.4.0
# sentence-transformers>=2.2.0

# Optional: Development and testing
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
